@@ -8,6 +8,6 @@ class LinqTest extends TestCaseEnumerable
 {
     function testFunctions()
     {
-        $this->assertInstanceOf('YaLinqo\Enumerable', from(new \EmptyIterator));
+        $this->assertInstanceOf(\YaLinqo\Enumerable::class, from(new \EmptyIterator));
     }
 }
